refactor: 아고라 입장시 리렌더링 문제 해결을 위해, 부모 컴포넌트에서 사용하는 state 분리
